refactoring; move conditional processing to separate file
